pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "4thGRC"
version = "0.1.0"
description = "4th.GRC™ — Agentic AI Governance Platform by Fourth Industrial Systems"
authors = [
    { name="Dr. Freeman A. Jackson", email="info@4th.is" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "streamlit>=1.35",
    "pytest>=8.2",
    "azure-identity>=1.16",
    "azure-keyvault-secrets>=4.8",
    "azure-storage-blob>=12.19",
    "azure-mgmt-resource>=23.0",
    "azure-cosmos>=4.5",
    "python-dotenv>=1.0",
    "opencensus-ext-azure>=1.1",
]

[project.urls]
Homepage = "https://4th.is"
Documentation = "https://learn.4th.is"
Source = "https://github.com/FourthIndustrialSystems/4th.GRC"
Issues = "https://github.com/FourthIndustrialSystems/4th.GRC/issues"

[tool.setuptools.packages.find]
where = ["policyengine", "services", "agents", "apps"]
include = ["policyengine*", "services*", "agents*", "apps*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
